fix host/build/target mix-up
Since the package to be created is for cross-compilation, os/arch must
actually specify the platform on which we use the package and not the
target platform. That latter should be specified via
os_target/arch_target instead. os_build/arch_build is not really
relevant, since we can also build a package for MacOS on Linux or
Windows.

Since conan's default settings.yml doesn't specify a os_target.api_level
for Android ATM, the API level is instead specified via the option
target_api_level.

Remove dependency and checks on setting compiler, since it applies to
the build platform, i.e. where we build the package, not where we use
it. Introduce an option target_libcxx accordingly.